ch6/src/main/java/ch6/mock/SAPInvoiceSender.java
package ch6.mock;

import ch6.stub.Invoice;
import ch6.stub.InvoiceFilter;

import java.util.List;

public class SAPInvoiceSender {

    private final InvoiceFilter filter;
    private final SAP sap;

    public SAPInvoiceSender(InvoiceFilter filter, SAP sap) {
        this.filter = filter;
        this.sap = sap;
    }

    public void sendLowValuedInvoices() {
        List<Invoice> lowValuedInvoices = filter.lowValueInvoices();
        for(Invoice invoice : lowValuedInvoices) {
             sap.send(invoice);
        }
    }
}
